pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redteam-composer"
version = "0.1.0"
description = "A CLI tool for browsing jailbreak techniques and composing instructions for AI red teaming"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
rtc = "redteam_composer.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
redteam_composer = ["taxonomy/**/*.yaml"]
